Remove pending UA shadow tree state when an input element is removed from the document https://bugs.webkit.org/show_bug.cgi?id=238429 <rdar://90682918>
Reviewed by Antti Koivisto.

Source/WebCore:

If an <input> element has a pending UA shadow tree, and the element is
removed from the document, we don't currently remove the element from
Document::m_elementsWithPendingUserAgentShadowTreeUpdates. Not only
does that means we may wastefully create the shadow tree if the
element is not inserted into the document again, but if the element is
then inserted into a different document, flushing style in the new
document won't correctly create the shadow tree.

Test: fast/forms/lazy-shadow-tree-creation-move-document-crash.html

* html/HTMLInputElement.cpp:
(WebCore::HTMLInputElement::removedFromAncestor):

LayoutTests:

* fast/forms/lazy-shadow-tree-creation-move-document-crash-expected.txt: Added.
* fast/forms/lazy-shadow-tree-creation-move-document-crash.html: Added.



Canonical link: https://commits.webkit.org/248926@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@291961 268f45cc-cd09-0410-ab3c-d52691b4dbfc