REGRESSION (r291587): [ iOS ] editing/deleting/insert-in-orphaned-selection-crash.html is a constant text failure (238282) https://bugs.webkit.org/show_bug.cgi?id=238282 <rdar://problem/90720306>
Reviewed by Aditya Keerthi.

In r291587 we started dropping native appearance for non-form control elements with
specified author style. This is the correct behavior, but requires adjusting this
test which relies on the appearance of a table element in order to reproduce a
non-appearance related crash. This patch updates the test to add table data tags which
ensure the table element gets included in the selectAll() command to keep the test passing.

* editing/deleting/insert-in-orphaned-selection-crash.html:
* platform/ios-wk2/TestExpectations:



Canonical link: https://commits.webkit.org/248943@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@291984 268f45cc-cd09-0410-ab3c-d52691b4dbfc