REGRESSION (r290794): [ iOS Debug ] 4 editing/selection/* tests are constant timeouts https://bugs.webkit.org/show_bug.cgi?id=238155 rdar://90593957
Reviewed by Aditya Keerthi.

Source/WebKit:

After the changes in r290794, we now preemptively compute and send an excessive amount of autocorrection
contexts to the UI process when an element gains focus, such that the UI process is capable of immediately
responding to calls to `-requestAutocorrectionContextWithCompletionHandler:` without relying on synchronous IPC
to the web process. This made several (already-fairly-slow) tests in editing/selection time out, due to WebKit
computing and sending an excessive amount of this contextual information.

To mitigate this, only perform this preemptive autocorrection context computation if the focused element is
actually changing (i.e., we're not refocusing the same element). Covered by removing several [ Slow ] test
expectations that should no longer be needed.

* WebProcess/WebPage/WebPage.cpp:
(WebKit::WebPage::elementDidFocus):

LayoutTests:

Unmark several tests that no longer need to be marked as [ Slow ].

* platform/ios/TestExpectations:



Canonical link: https://commits.webkit.org/248955@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@292005 268f45cc-cd09-0410-ab3c-d52691b4dbfc