HTMLAttachmentElement.getAttachmentIdentifier() should propagate attachment data to the client https://bugs.webkit.org/show_bug.cgi?id=238473 rdar://90938796
Reviewed by Devin Rousso.

Source/WebCore:

In the case where a WebKit client that has enabled the attachment element uses
`HTMLAttachmentElement.getAttachmentIdentifier` on an image element to create a new attachment-backed image, we
currently propagate an empty `_WKAttachment` to the UI delegate in the UI process, even in the case where the
image element has a visible image that has been loaded.

To support plain-text editing mode, Mail needs to be able to call `getAttachmentIdentifier` to create attachment
elements for images that already contain images with blob sources, but are inserted into the document via
JavaScript (as opposed to native editing code).

To make this work, we plumb the image element itself through the `registerAttachmentIdentifier` method that
currently only takes a single attachment identifier, and use it to send the image data, image name, and MIME
type to the attachment client if the image has a loaded image; otherwise, we fall back to creating and surfacing
an empty _WKAttachment to the client.

Test: WKAttachmentTests.CreateAttachmentsFromExistingImage

* dom/Document.cpp:
(WebCore::Document::registerAttachmentIdentifier):
* dom/Document.h:
* editing/Editor.cpp:
(WebCore::Editor::registerAttachmentIdentifier):
* editing/Editor.h:
* html/HTMLAttachmentElement.cpp:
(WebCore::HTMLAttachmentElement::getAttachmentIdentifier):

Tools:

Add a new API test to exercise the fix.

* TestWebKitAPI/Tests/WebKitCocoa/WKAttachmentTests.mm:
(TestWebKitAPI::TEST):



Canonical link: https://commits.webkit.org/248975@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@292032 268f45cc-cd09-0410-ab3c-d52691b4dbfc