Navigation Timing data is corrupt in WebView (UIWebView/WKWebView) https://bugs.webkit.org/show_bug.cgi?id=186919 <rdar://41393423>
Reviewed by Simon Fraser.

Source/WebCore:

Sometimes the observed requestStart time is after the observed responseStart time.
This may be due to the device's clock changing.  It may be due to something else,
but currently CFNetwork does not use a monotonic timer to gather the data.
When this happens, use the later of the two times for responseStart.

* platform/network/cocoa/NetworkLoadMetrics.mm:
(WebCore::packageTimingData):

Source/WebKit:

* NetworkProcess/cocoa/NetworkSessionCocoa.mm:
(-[WKNetworkSessionDelegate URLSession:task:didFinishCollectingMetrics:]):



Canonical link: https://commits.webkit.org/245886.379@safari-613-branch
git-svn-id: https://svn.webkit.org/repository/webkit/branches/safari-613-branch@292055 268f45cc-cd09-0410-ab3c-d52691b4dbfc