Web Inspector: Add check for null `entryScope` in JSC::Debugger::detach https://bugs.webkit.org/show_bug.cgi?id=238518
Reviewed by Devin Rousso.

A virtual machine may not always have an `entryScope`, which can occur if a `JSGlobalObject` is being destructed
(which will call into `JSC::Debugger::Detach`) before a `VMEntryScope` is created or after a `VMEntryScope` has
fallen out of scope, causing the VM's `entryScope` to be null again. In the original internal issue that led to
finding this, this most commonly occurred in conjunction with page refresh while the debugger was paused.

Because it is possible that we were previously paused in the VM that now has no `entryScope`, we should go ahead
and unpause.

* debugger/Debugger.cpp:
(JSC::Debugger::detach):



Canonical link: https://commits.webkit.org/249004@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@292075 268f45cc-cd09-0410-ab3c-d52691b4dbfc