Split streaming of AX trees from the class AXLogger. https://bugs.webkit.org/show_bug.cgi?id=238550 <rdar://problem/91041921>
Reviewed by Chris Fleizach.

AXLogger functionality is debug only. We need to stream the content of
the AX trees both in debug and release builds. This patch splits
streaming of the AX trees from the AXLogger class, which remains a debug
only utility.
In addition, added a mechanism to control what is streamed via the
AXStreamOptions enumeration.

* accessibility/AXLogger.cpp:
(WebCore::shouldLog):
(WebCore::operator<<):
(WebCore::streamIsolatedSubtreeOnMainThread):
Streams a snapshot of the hierarchy of AXIsolatedObjects on the main
thread, limiting the properties of the AXIsolatedObject that can be
accessed using AXStreamOptions.

(WebCore::streamAXCoreObject):
(WebCore::streamSubtree):
(WebCore::AXLogger::add): Replaced with streamSubtree.
* accessibility/AXLogger.h:
* accessibility/isolatedtree/AXIsolatedTree.h:



Canonical link: https://commits.webkit.org/249049@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@292142 268f45cc-cd09-0410-ab3c-d52691b4dbfc