Remove push subscriptions when permissions are reset https://bugs.webkit.org/show_bug.cgi?id=238068
Reviewed by Youenn Fablet.

Source/WebKit:

When the notification permission for an origin is reset (i.e. is in the prompt state), we
should delete any push subscriptions associated with that origin. We do this by having the
providerDidRemoveNotificationPolicies API call deletePushAndNotificationRegistration in
webpushd.

We also do this when processing a push message, since it's possible for the state to get out
of sync due to the number of processes involved. For instance, removing the policy from
System Preferences rather than the browser's notification preference pane can result in the
providerDidRemoveNotificationPolicies call in the browser to be delayed or dropped.

Note that we don't delete the subscription if the permission is simply toggled off (i.e. the
permission is in the Denied state). In that case, the user can easily toggle the permission
back on. We want the subscription object to stay the same after the preference is toggled
back on so that the server can still send pushes to that push endpoint. In a future patch
we'll add subscriptions in this state to the ignored topics list in webpushd, which will
prevent pushes for those subscriptions from being delivered to the device without actually
unsubscribing.

* NetworkProcess/NetworkProcess.cpp:
(WebKit::NetworkProcess::processPushMessage):
* NetworkProcess/NetworkProcess.h:
* NetworkProcess/NetworkProcess.messages.in:
* UIProcess/Network/NetworkProcessProxy.cpp:
(WebKit::NetworkProcessProxy::processPushMessage):
* UIProcess/Notifications/WebNotificationManagerProxy.cpp:
(WebKit::removePushSubscriptionsForOrigins):
(WebKit::WebNotificationManagerProxy::providerDidRemoveNotificationPolicies):
* webpushd/WebPushDaemon.mm:
(WebPushD::Daemon::deletePushAndNotificationRegistration):

Tools:

Add a test to make sure push subscriptions are removed when the notification policy for a
website is reset.

Also fixed an issue with a PushAPI test that didn't properly register for notification permissions.

* TestWebKitAPI/TestNotificationProvider.cpp:
(TestWebKitAPI::TestNotificationProvider::TestNotificationProvider):
(TestWebKitAPI::TestNotificationProvider::notificationPermissions const):
(TestWebKitAPI::TestNotificationProvider::setPermission):
(TestWebKitAPI::TestNotificationProvider::resetPermission):
* TestWebKitAPI/TestNotificationProvider.h:
* TestWebKitAPI/Tests/WebKitCocoa/PushAPI.mm:
* TestWebKitAPI/Tests/WebKitCocoa/WebPushDaemon.mm:

LayoutTests:

Fix an issue with a layout test that didn't register for notification permissions.

* http/wpt/service-workers/service-worker-spinning-push.https.html:



Canonical link: https://commits.webkit.org/249065@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@292158 268f45cc-cd09-0410-ab3c-d52691b4dbfc