[Bugzilla] Code reviews show non-ASCII characters in patches as garbage https://bugs.webkit.org/show_bug.cgi?id=238630
Reviewed by Aakash Jain and Darin Adler.

We have a convention that our sources (and therefore our patches) in WebKit are UTF-8.
The output of prettify.rb is also UTF-8. So, we just need to tell Perl that, when it
reads the output of prettify.rb, it just needs to understand that the data it's reading
is UTF-8.

* attachment.cgi:
(prettyPatch):


Canonical link: https://commits.webkit.org/249077@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@292170 268f45cc-cd09-0410-ab3c-d52691b4dbfc