Service-Worker-Navigation-Preload header not being sent when Navigation Preload is enabled. https://bugs.webkit.org/show_bug.cgi?id=238564
Reviewed by Alex Christensen.

Source/WebKit:

We were cancelling the preload as soon as receiving a response through FetchEvent.respondWith.
But it is possible to answer with a synthetic response and then use the preload to fill the synthetic response body.
To allow this, we no longer cancel the preload when receiving a response in case of enabled navigation preload.

Test: http/wpt/service-workers/service-worker-iframe-preload-after-response.https.html

* NetworkProcess/ServiceWorker/ServiceWorkerFetchTask.cpp:

LayoutTests:

* http/wpt/service-workers/resources/service-worker-iframe-preload-after-response-script.py: Added.
* http/wpt/service-workers/service-worker-iframe-preload-after-response-worker.js: Added.
* http/wpt/service-workers/service-worker-iframe-preload-after-response.https-expected.txt: Added.
* http/wpt/service-workers/service-worker-iframe-preload-after-response.https.html: Added.



Canonical link: https://commits.webkit.org/249194@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@292296 268f45cc-cd09-0410-ab3c-d52691b4dbfc